Added quoted symbols
So you can now say things like 'foo bar' to get a symbol with a space in
it.  Unlike prolog, this won't work for treating capitalized symbols as
constants rather than variables (sorry!).  This was added so that it's
possible to call internal things like %call_failed/2 from user code if
you really want to.