Added support for calling predicates from C# code with arguments.
IsTrue and RunBotL now take a params argument with arguments to pass to
the predicate.  Note that this means any integer or float arguments will
get boxed, and the whole thing will be packaged as an array.  So using
this feature allocates memory.  If you want to avoid allocating memory,
store arguments into GlobalVariable objects and call using the old,
argument-free interface, and that will run with no dynamic allocate.

Also, there isn't currently a facility for getting an answer back, e.g.
like UnityProlog's SolveFor.  That will come next.